Implement HashStable for Span in libsyntax_pos.
